LPS-3814 Source formatting
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@33353 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portlet/assettagsnavigation/action/ConfigurationActionImpl.java
@@ -54,7 +54,7 @@ public class ConfigurationActionImpl extends BaseConfigurationAction {
 			return;
 		}
 
-		int classNameId = ParamUtil.getInteger(actionRequest, "classNameId");
+		long classNameId = ParamUtil.getLong(actionRequest, "classNameId");
 		String displayStyle = ParamUtil.getString(
 			actionRequest, "displayStyle");
 		boolean showAssetCount = ParamUtil.getBoolean(
